[Editorial] her's -> hers
The `her's` spelling can be seen in, e.g.,
https://archive.org/details/in.ernet.dli.2015.46179/page/n125/mode/2up,
but not in e.g.,
https://babel.hathitrust.org/cgi/pt?id=coo.31924013208453&view=1up&seq=225. I
checked a handful of other scans on Internet Archive, and they all
have `hers`.

Looking on Project Gutenberg,
https://www.gutenberg.org/files/1342/1342-h/1342-h.htm#link2HCH0022
has `hers`, while, e.g.,
https://www.gutenberg.org/files/42671/42671-h/42671-h.htm has `her's`.

According to
[Wiktionary](https://en.wiktionary.org/wiki/her%27s#English), `her's`
is now non-standard, so it may well have been present in the
original. https://www.grammarphobia.com/blog/2010/09/his-and-hers.html
suggests that `hers` would have come into standard usage sometime in the
1500s, though.

At any rate, `her's` definitely looks out of place to modern
readers. There are also no other instances of `her's` in the text, but
a handful of instances of `hers` (as determined with `git grep
--word-regexp "her's"` and `git grep --word-regexp "hers"`).